+++ packages/api/src/controllers/v2/mod.rs
@@ -1,8 +1,9 @@
 use deliberations::_id::responses::DeliberationResponseController;
 use models::*;
+use reviews::ReviewControllerV2;
 use surveys::_id::responses::SurveyResponseController;
 
-use super::{resources::v1::bucket::MetadataControllerV1, reviews::v1::ReviewControllerV1};
+use super::resources::v1::bucket::MetadataControllerV1;
 
 pub mod surveys {
     pub mod _id {
@@ -20,6 +21,8 @@ pub mod deliberations {
     }
 }
 
+pub mod reviews;
+
 #[derive(Clone, Debug)]
 pub struct Version2Controller {}
 
@@ -34,7 +37,7 @@ impl Version2Controller {
                 "/organizations",
                 crate::controllers::organizations::v2::OrganizationController::route(pool.clone())?,
             )
-            .nest("/reviews", ReviewControllerV1::route(pool.clone())?)
+            .nest("/reviews", ReviewControllerV2::route(pool.clone())?)
             .nest("/metadata", MetadataControllerV1::route(pool.clone())?)
             .nest(
                 "/deliberations/:deliberation-id/responses",
